Share plan-aware billing list with full page
