+++ app/[locale]/dashboard/billing/billing-client.tsx
@@ -1,24 +1,35 @@
 'use client'
 
-import { Card, CardContent } from "@/components/ui/card"
+import { useI18n } from '@/locales/client'
 import BillingManagement from './components/billing-management'
+import { BillingPlanList } from './components/billing-plan-list'
 
 export default function BillingPage() {
+  const t = useI18n()
+
   return (
-    <div className="flex w-full relative min-h-screen">
-      <div className='flex flex-1 w-full'>
-        <div className={`w-full flex flex-col lg:flex-row min-h-screen`}>
-          <main className="w-full py-6 lg:py-8">
-            <div className="container mx-auto px-4">
-              <Card className="border-none shadow-none bg-transparent">
-                <CardContent className="p-0">
-                      <BillingManagement />
-                </CardContent>
-              </Card>
-            </div>
-          </main>
+    <main className="min-h-[calc(100dvh-var(--navbar-height,4rem))] bg-[#FAFAFA] px-4 py-8 dark:bg-background sm:px-6 lg:px-10 lg:py-12">
+      <div className="mx-auto w-full max-w-3xl">
+        <header className="mb-6">
+          <h1 className="text-2xl font-semibold tracking-[-0.03em] text-[#171717] dark:text-foreground sm:text-3xl">
+            {t('dashboard.billingSheet.title')}
+          </h1>
+          <p className="mt-1 text-sm text-[#686D67] dark:text-muted-foreground">
+            {t('dashboard.billingSheet.description')}
+          </p>
+        </header>
+
+        <section className="overflow-hidden rounded-[4px] border border-[#E5E5E5] bg-white dark:border-border dark:bg-card">
+          <BillingPlanList
+            contentClassName="p-4 sm:p-6"
+            footerClassName="p-4 sm:p-6"
+          />
+        </section>
+
+        <div className="mt-8">
+          <BillingManagement showPlanSections={false} />
         </div>
       </div>
-    </div>
+    </main>
   )
 }
